chore: split agent graph empty state renderer
Refs #692
